fips203

Module traits

All functionality is covered by traits, such that consumers can utilize trait objects if desired.

Traits

Decaps

The Decaps trait defines methods for recovering the shared secret using a decapsulation key and ciphertext.

Encaps

The Encaps trait defines methods for generating shared secrets and ciphertexts using an encapsulation key.

KeyGen

The KeyGen trait is defined to allow trait objects.

SerDes

The SerDes trait provides methods for serializing and deserializing cryptographic objects.